v1/separate files for common params
